+++ bank_account2/NormalAccount.h
@@ -24,7 +24,7 @@ public:
 
 	virtual void SetMoney(int money)   //입금
 	{
-		Account::SetMoney(money + money*interestRatio*0.01);
+		Account::SetMoney(money + int(money*interestRatio*0.01));
 		//입금 + (입금 * 이자율)
 	}
 };
